Added CMAKE support
Added support for the CMake build system (C/C++). It uses the cmake-build-debug directory as it's build directory. This is a useful system to support, as this directory can grow to massive sizes within large projects.

Support was added by adding an additional subcommand, in the same manner as with `node_modules` and `target`. A corresponding arm in a match statement was also made.

This change was tested on an up-to-date build of Manjaro Linux.